embeddings_kd/requirements.txt
albumentations
clearml
fire
matplotlib
numpy
opencv-python
pandas
h5py
Pillow
lightning[extra]
torch
torchvision
tqdm
timm
git+https://github.com/deepglint/unicom.git